refactor: Extract server to own package for easier reuse
This could then be used in other projects.
