+++ capi-runtime-extensions.code-workspace
@@ -1,0 +1,16 @@
+// Copyright 2021 D2iQ, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+{
+    "folders": [
+        {
+            "path": "."
+        },
+        {
+            "path": "./server"
+        }
+    ],
+    "settings": {
+        "go.buildTags": "e2e"
+    }
+}
