+++ cmd/capi-runtime-extensions/main.go
@@ -11,15 +11,24 @@ import (
 
 	"github.com/spf13/pflag"
 	"golang.org/x/sync/errgroup"
+	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	cliflag "k8s.io/component-base/cli/flag"
 	"k8s.io/component-base/logs"
 	logsv1 "k8s.io/component-base/logs/api/v1"
 	"k8s.io/component-base/version/verflag"
 	"k8s.io/klog/v2"
+	capiv1 "sigs.k8s.io/cluster-api/api/v1beta1"
+	crsv1 "sigs.k8s.io/cluster-api/exp/addons/api/v1beta1"
 	ctrl "sigs.k8s.io/controller-runtime"
+	ctrclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/d2iq-labs/capi-runtime-extensions/internal/controllermanager"
-	runtimewebhooks "github.com/d2iq-labs/capi-runtime-extensions/internal/runtimehooks/webhooks"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/cni/calico"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/httpproxy"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/servicelbgc"
+	"github.com/d2iq-labs/capi-runtime-extensions/server/pkg/server"
 )
 
 var (
@@ -43,13 +52,40 @@ func main() {
 	// Creates a logger to be used during the main func.
 	setupLog := ctrl.Log.WithName("main")
 
-	runtimeWebhookServer := runtimewebhooks.NewServer()
 	controllers := controllermanager.New()
+
+	scheme := runtime.NewScheme()
+	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
+	utilruntime.Must(crsv1.AddToScheme(scheme))
+	utilruntime.Must(capiv1.AddToScheme(scheme))
+
+	// Gets a client to access the Kubernetes cluster where this RuntimeExtension will be deployed to
+	restConfig, err := ctrl.GetConfig()
+	if err != nil {
+		setupLog.Error(err, "error getting config for the cluster")
+		os.Exit(1)
+	}
+
+	client, err := ctrclient.New(restConfig, ctrclient.Options{Scheme: scheme})
+	if err != nil {
+		setupLog.Error(err, "error creating client to the cluster")
+		os.Exit(1)
+	}
+
+	calicoCNIConfig := &calico.CalicoCNIConfig{}
+
+	runtimeWebhookServer := server.NewServer(
+		servicelbgc.New(client),
+		calico.New(client, calicoCNIConfig),
+		httpproxy.NewVariable(),
+		httpproxy.NewPatch(),
+	)
 
 	// Initialize and parse command line flags.
 	initFlags(pflag.CommandLine)
 	runtimeWebhookServer.AddFlags("runtimehooks", pflag.CommandLine)
 	controllers.AddFlags("controllermanager", pflag.CommandLine)
+	calicoCNIConfig.AddFlags("runtimehooks.calicocni", pflag.CommandLine)
 	pflag.CommandLine.SetNormalizeFunc(cliflag.WordSepNormalizeFunc)
 	pflag.CommandLine.AddGoFlagSet(flag.CommandLine)
 	pflag.Parse()
