+++ pkg/handlers/auditpolicy/inject.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/d2iq-labs/capi-runtime-extensions/pkg/capi/clustertopology/patches"
 	"github.com/d2iq-labs/capi-runtime-extensions/pkg/capi/clustertopology/patches/selectors"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers"
+	"github.com/d2iq-labs/capi-runtime-extensions/server/pkg/handlers"
 )
 
 const (
